Extract presale logic into reusable components
